commune/modules/polaris/polaris/compute_subnet/requirements.txt
fastapi
uvicorn
python-dotenv
docker
pydantic
python-multipart
pytest
pytest-cov